server/pyproject.toml
[project]
name = "a2a-server"
version = "0.1.0"
description = "Agent2Agent  server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pydantic>=2.10.6",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "mesop>=1.0.0",
    "a2a-samples",
    "pandas>=2.2.0",
    "google-genai>=1.9.0",
    "google-adk>=0.0.3",
    "pymysql",
    "python-dotenv",
    "solana>=0.32.0",
    "solders>=0.22.0",
    "PyNaCl>=1.5.0",
    "pyyaml>=6.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["a2a_ui"]

[tool.uv.sources]
a2a-samples = { path = "../../samples/python", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
